move components into subfolders
