+++ client/src/components/index.ts
@@ -1,10 +1,10 @@
-export * from './AuditorCount';
-export * from './CurrentTrack';
-export * from './Jingle';
-export * from './Like';
-export * from './Player';
-export * from './SideMenu';
-export * from './SettingModal';
-export * from './SongModal';
-export * from './TopMenu';
-export * from './UserModal';
+export * from './now/AuditorCount';
+export * from './now/CurrentTrack';
+export * from './jingle/Jingle';
+export * from './interaction/Like';
+export * from './interaction/Player';
+export * from './interaction/SideMenu';
+export * from './modal/SettingModal';
+export * from './modal/SongModal';
+export * from './interaction/TopMenu';
+export * from './modal/UserModal';
